ai-backend/requirements.txt
opencv-python
mediapipe
streamlit
pillow
